Fix test humanoid normal attacks not connecting normals 2-4
Assign the correct data in the AIState you idiot

+++ ARPGProject/Assets/Scripts/UI/PlayerHudUI.cs
@@ -22,7 +22,7 @@ namespace Bebis {
         }
 
         private void OnCurrentHealthChanged(int newHealth) {
-            _playerHealthBar.SetFillPercent(newHealth / PlayerCharacter.Instance.Damageable.MaxHealth);
+            _playerHealthBar.SetFillPercent((float)newHealth / PlayerCharacter.Instance.Damageable.MaxHealth);
             _healthTextDisplay.text = $"{newHealth} / {PlayerCharacter.Instance.Damageable.MaxHealth}";
         }
